requirements.txt
torch
torchvision
tqdm 
matplotlib
pyyaml